# services/jurisdiction-rules-svc/openapi.yaml
openapi: "3.1.0"
info:
  title: Jurisdiction Rules Service
  version: "1.1.0"
  description: |
    Authoritative registry of jurisdiction identities and their effective-dated
    applicability rules (03-microservices.md §8.2).

    **Governance invariants:**
    - `rule_payload` holds APPLICABILITY METADATA ONLY — never computation values.
      Thresholds, rates and bands belong to the Tax and Payroll services (OQ-1, Model B).
    - Every type/status/domain field is a VARCHAR tag stored as data. New jurisdiction
      types, rule domains and status values are added by data migration, never by a
      code change (OQ-3).
    - Nothing is deleted. Jurisdictions are deactivated with `active_flag` + `effective_to`;
      rules are end-dated and moved to SUPERSEDED or RETIRED.
    - Historical actions must always be explainable against the rule set active at the
      time of execution, so point-in-time reads keep working after a jurisdiction is
      deactivated and after a rule is superseded.
    - `legal_drift_state` transitions are append-only. The column carries the current
      value; `/v1/rules/{id}/drift-events` carries the history (OQ-4).
    - Jurisdiction and rule data are platform-wide reference data. There is no
      `tenant_id` and no row-level security on these tables.

    **Read vs runtime resolution.** `/rules` is the raw per-jurisdiction, effective-dated
    view — the audit and replay surface, one jurisdiction only. `/rule-pack` is the
...

...

...
servers:
...
paths:
...
  /v1/rules/{jurisdiction_rule_id}:
    get:
      tags: [Rules]
      summary: Fetch one rule by id
      description: |
        The rule id appears in every published event and in the rule basis recorded
        against governed actions, so it has to be resolvable directly.
      operationId: getRule
      parameters:
        - $ref: '#/components/parameters/RuleID'
        - $ref: '#/components/parameters/CorrelationID'
      responses:
        "200":
          description: The rule, in whatever status it currently holds.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JurisdictionRule' }
        "404": { $ref: '#/components/responses/NotFound' }
        "503": { $ref: '#/components/responses/StoreUnavailable' }
...
  /v1/admin/jurisdictions/{jurisdiction_id}/rules:
    post:
      tags: [Admin, Rules]
      summary: Create a rule
      description: |
        Idempotent on the dedup key (`jurisdiction_id`, `rule_code`, `effective_from`).

        A rule may only be created in DRAFT (the default) or ACTIVE. SUPERSEDED and
        RETIRED are reached through the transition endpoint, never directly.

        Two live rules with the same `rule_code` and overlapping effective periods are
        rejected with `409 overlapping_rule`, because both would satisfy the same
        point-in-time query. DRAFT rules are exempt so a replacement can be prepared
        while the incumbent is still in force — the same check runs again on activation.
      operationId: createRule
      security: [{ principalHeader: [] }]
      parameters:
        - $ref: '#/components/parameters/JurisdictionID'
        - $ref: '#/components/parameters/PrincipalID'
        - $ref: '#/components/parameters/CorrelationID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRuleRequest' }
      responses:
        "201":
          description: |
            Created. Publishes `jurisdiction.rule.updated`, plus
            `jurisdiction.rule.activated` when created directly in ACTIVE.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JurisdictionRule' }
        "200":
          description: Idempotent replay of an existing rule. Publishes nothing.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JurisdictionRule' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/MissingPrincipal' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404":
          description: |
            `jurisdiction_not_found` — the jurisdiction is unknown or has been deactivated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        "409":
          description: |
            `conflict` — dedup key matches an existing rule with a different payload or
            name. `overlapping_rule` — the effective period overlaps another live rule
            with the same `rule_code`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        "413": { $ref: '#/components/responses/BodyTooLarge' }
        "503": { $ref: '#/components/responses/AuthzOrStoreUnavailable' }
...
components:
  securitySchemes:
    principalHeader:
      type: apiKey
      in: header
      name: X-Principal-Id
      description: |
        The acting principal, set by the gateway's ForwardAuth middleware after it
        verifies the caller's identity envelope against identity-svc's JWKS. Callers
        do not set this themselves — a request that reaches the service without it is
        rejected with `401 missing_principal` rather than attributed to the platform.
  parameters:
    JurisdictionID:
      name: jurisdiction_id
      in: path
      required: true
      description: Jurisdiction UUID. A malformed value answers 404, not 503.
      schema: { type: string }
    RuleID:
      name: jurisdiction_rule_id
      in: path
      required: true
      description: Rule UUID. A malformed value answers 404, not 503.
      schema: { type: string }
    CorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      description: |
        Evidence propagation id, echoed on every response. Generated by the service
        when absent.
      schema: { type: string }
    PrincipalID:
      name: X-Principal-Id
      in: header
      required: true
      description: See the `principalHeader` security scheme.
      schema: { type: string }
...
  responses:
    BadRequest:
      description: |
        Malformed or incomplete request. Codes include `missing_field`,
        `invalid_request_body`, `invalid_effective_at`, `invalid_effective_period`,
        `invalid_limit`, `invalid_offset`, `invalid_rule_status`,
        `invalid_rule_payload`, `invalid_status`, `invalid_drift_state`.
        An unknown JSON field is rejected rather than ignored.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    MissingPrincipal:
      description: '`missing_principal` — no caller identity on the request.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: '`unauthorized` — the Authorization Service denied the action.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: |
        `jurisdiction_not_found` / `rule_not_found`. Also returned for a malformed id,
        which cannot name an existing row.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BodyTooLarge:
      description: '`request_body_too_large` — the body exceeds 256 KiB.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    StoreUnavailable:
      description: |
        `store_unavailable` — the service cannot answer. This is NOT "no". Callers
        validating a `jurisdiction_id` must reject the write fail-closed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    AuthzOrStoreUnavailable:
      description: |
        `authz_unavailable` — no authorization decision could be obtained, so the
        mutation was refused. `store_unavailable` — the database could not be reached.
        Both fail closed: nothing was written.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [type, title, status, detail, instance]
      description: |
        RFC 9457 Problem Details format. Clients can switch on `type` or `title`.
        The `extension` fields provide machine-readable codes.
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type.
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          description: HTTP status code generated by the origin server.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence.
        code:
          type: string
          description: Stable machine-readable code for client logic (e.g., "missing_field").
        field:
          type: string
          description: The offending field, when code is "missing_field".
        jurisdiction_id:
          type: string
          format: uuid
          description: The jurisdiction ID, when relevant.
...
    JurisdictionRule:
      type: object
      description: |
        An effective-dated applicability rule. INTERNAL data
        (data_classification_audit.md §2.11).
      required:
        - jurisdiction_rule_id
        - jurisdiction_id
        - rule_domain
        - rule_code
        - rule_name
        - effective_from
        - rule_payload
        - rule_status
        - legal_drift_state
        - data_classification
        - created_at
        - created_by_principal_id
        - schema_version
      properties:
        jurisdiction_rule_id: { type: string, format: uuid }
        jurisdiction_id: { type: string, format: uuid }
        rule_domain:
          type: string
          description: e.g. PAYROLL, TAX, EMPLOYMENT, FILING, RETENTION, BENEFITS. Data driven.
        rule_code:
          type: string
          description: Stable machine-readable identifier for this rule type within a domain.
        rule_name: { type: string }
        effective_from: { type: string, format: date-time }
        effective_to:
          type: [string, "null"]
          format: date-time
          description: |
            Exclusive end of the half-open effective period. Set automatically when the
            rule is superseded or retired.
        rule_payload:
          type: object
          description: |
            APPLICABILITY METADATA ONLY — never computation values. Thresholds, rates
            and bands belong to the Tax and Payroll services (OQ-1, Model B). Structure
            is domain-specific and governed by `rule_domain` + `schema_version`.
            Example: `{"applies_to_entity_types":["COMPANY","BRANCH"],
            "filing_frequency":"MONTHLY","authority_code":"HMRC"}`.
        source_reference:
          type: [string, "null"]
          description: Citation — legislative act, regulation, statutory instrument.
        external_feed_reference:
          type: [string, "null"]
          description: The external regulatory data feed that produced this rule.
        rule_status:
          type: string
          description: DRAFT, ACTIVE, SUPERSEDED or RETIRED. VARCHAR, not an enum.
        rule_version:
          type: integer
          description: Immutable version number. Incremented on each status change.
        supersedes_rule_id:
          type: [string, "null"]
          format: uuid
          description: The rule this rule supersedes, forming a version chain.
        precedence_level:
          type: integer
          description: Explicit precedence level. Lower = higher precedence. Used for conflict resolution.
        legal_drift_state:
          type: string
          description: |
            CURRENT, DRIFTED or UNDER_REVIEW. Current value only — the transition
            history is at /v1/rules/{id}/drift-events.
        data_classification: { type: string, const: "INTERNAL" }
        created_at: { type: string, format: date-time }
        created_by_principal_id: { type: string }
        schema_version: { type: string }
        updated_at: { type: [string, "null"], format: date-time }
        updated_by_principal_id: { type: [string, "null"] }
...
    CreateRuleRequest:
      type: object
      required: [rule_domain, rule_code, rule_name, effective_from]
      additionalProperties: false
      properties:
        rule_domain: { type: string, minLength: 1, maxLength: 64 }
        rule_code: { type: string, minLength: 1, maxLength: 128 }
        rule_name: { type: string, minLength: 1 }
        effective_from: { type: string, format: date-time }
        effective_to:
          type: [string, "null"]
          format: date-time
          description: Must be strictly after effective_from.
        rule_payload:
          type: object
          description: |
            Applicability metadata. Must be a JSON object — a scalar or array is
            rejected. Omitted defaults to `{}`.
          default: {}
        source_reference: { type: [string, "null"] }
        external_feed_reference: { type: [string, "null"] }
        rule_status:
          type: string
          enum: [DRAFT, ACTIVE]
          default: DRAFT
          description: |
            SUPERSEDED and RETIRED are not accepted here — they are reached only
            through the transition endpoint.
        rule_version:
          type: integer
          default: 1
          description: Immutable version number. Defaults to 1 for new rules.
        supersedes_rule_id:
          type: [string, "null"]
          format: uuid
          description: The rule this rule supersedes, forming a version chain.
        precedence_level:
          type: integer
          default: 0
          description: Explicit precedence level. Lower = higher precedence.
...